requirements.txt
annotated-types==0.6.0
anyio==4.3.0
Authlib==1.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
databases==0.9.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
spotipy==2.23.0
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wheel==0.41.2
